+++ Samples/iOS-SwiftUI/iOS-SwiftUI/SwiftUIApp.swift
@@ -11,6 +11,7 @@ struct SwiftUIApp: App {
             options.tracesSampleRate = 1.0
             options.profilesSampleRate = 1.0
             options.experimental.sessionReplay.sessionSampleRate = 1.0
+            options.experimental.enableFileManagerSwizzling = true
             options.initialScope = { scope in
                 scope.injectGitInformation()
                 return scope